requirements.txt
faster-whisper==0.10.1
deepgram-sdk==2.14.1
pyannote.audio==3.2.1
openai==1.30.2
anthropic==0.18.2
opencv-python-headless==4.10.0.82
ffmpeg-python==0.2.0
rich==13.7.1
tqdm==4.66.4
psycopg2-binary==2.9.9
redis==5.0.4
prometheus-client==0.20.0
bottle==0.12.25
python-dotenv==1.0.1
sentry-sdk==2.5.0

# Production API dependencies
celery==5.3.4
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2